Fix nonlinear constraints for optqnips
Fixes a segfault when using nonlinear constraints.
Objects are made on the C++ side then passed by reference to python.
This is due to several of the C++ constructors wrapping the raw input
pointers in smart pointers causing objects to be deleted both by C++ and
python.
There is a decent chance that this introduces some memory leaks. This
should be looked into.